+++ Budget/CategoriasViewController.swift
@@ -1,0 +1,131 @@
+//
+//  CategoriasViewController.swift
+//  Budget
+//
+//  Created by md10 on 3/23/16.
+//  Copyright © 2016 Budget. All rights reserved.
+//
+
+import UIKit
+import CoreData
+
+class CategoriasViewController: UITableViewController {
+
+    var categoria: Categoria?
+    
+    @IBOutlet var labels: [UILabel]!
+    @IBOutlet weak var txtNome: UITextField!
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        if let categoria = categoria{
+            txtNome.text = categoria.nome!
+        }
+        
+    }
+    
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+        // Dispose of any resources that can be recreated.
+    }
+    
+    @IBAction func btnCancel(sender: AnyObject) {
+        
+        dissmissViewController()
+    }
+    
+    
+    @IBAction func btnSave(sender: AnyObject) {
+        
+
+        addConta()
+        navigationController?.popViewControllerAnimated(true)
+
+    }
+    
+    func dissmissViewController(){
+        navigationController?.popToRootViewControllerAnimated(true)
+    }
+    
+    func addConta(){
+        
+        categoria = Categoria.getCategoria()
+        
+        categoria?.nome = txtNome.text
+        do{
+            try categoria?.managedObjectContext?.save()
+        }catch{
+            let alert = Notification.mostrarErro("Desculpe", mensagem: "Não foi possível registrar")
+            presentViewController(alert, animated: true, completion: nil)
+        }
+    }
+
+    // MARK: - Table view data source
+
+    override func numberOfSectionsInTableView(tableView: UITableView) -> Int {
+        // #warning Incomplete implementation, return the number of sections
+        return 1
+    }
+
+    override func tableView(tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        // #warning Incomplete implementation, return the number of rows
+        return 1
+    }
+
+    /*
+    override func tableView(tableView: UITableView, cellForRowAtIndexPath indexPath: NSIndexPath) -> UITableViewCell {
+        let cell = tableView.dequeueReusableCellWithIdentifier("reuseIdentifier", forIndexPath: indexPath)
+
+        // Configure the cell...
+
+        return cell
+    }
+    */
+
+    /*
+    // Override to support conditional editing of the table view.
+    override func tableView(tableView: UITableView, canEditRowAtIndexPath indexPath: NSIndexPath) -> Bool {
+        // Return false if you do not want the specified item to be editable.
+        return true
+    }
+    */
+
+    /*
+    // Override to support editing the table view.
+    override func tableView(tableView: UITableView, commitEditingStyle editingStyle: UITableViewCellEditingStyle, forRowAtIndexPath indexPath: NSIndexPath) {
+        if editingStyle == .Delete {
+            // Delete the row from the data source
+            tableView.deleteRowsAtIndexPaths([indexPath], withRowAnimation: .Fade)
+        } else if editingStyle == .Insert {
+            // Create a new instance of the appropriate class, insert it into the array, and add a new row to the table view
+        }    
+    }
+    */
+
+    /*
+    // Override to support rearranging the table view.
+    override func tableView(tableView: UITableView, moveRowAtIndexPath fromIndexPath: NSIndexPath, toIndexPath: NSIndexPath) {
+
+    }
+    */
+
+    /*
+    // Override to support conditional rearranging of the table view.
+    override func tableView(tableView: UITableView, canMoveRowAtIndexPath indexPath: NSIndexPath) -> Bool {
+        // Return false if you do not want the item to be re-orderable.
+        return true
+    }
+    */
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepareForSegue(segue: UIStoryboardSegue, sender: AnyObject?) {
+        // Get the new view controller using segue.destinationViewController.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
