+++ Budget/Conta+CoreDataProperties.swift
