+++ Budget/Receita+CoreDataProperties.swift
